microservices/db_service/requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
sqlalchemy